found the minus, and ran1 included

+++ Project4/include/spin_initializer.h
@@ -3,6 +3,6 @@
 #include <time.h>
 #include <stdlib.h>
 
-void Initialize_spins(int** spin_matrix, int L, bool order, int& magnetization, int& energy);
+void Initialize_spins(int** spin_matrix, int L, bool order, double& magnetization, double& energy);
 
 #endif // SPIN_INITIALIZER_H
